% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{apply_inverse_transformations}
\alias{apply_inverse_transformations}
\title{Apply Inverse Transformations}
\usage{
apply_inverse_transformations(sim, wt, parm, names)
}
\arguments{
\item{sim}{Simulated weather data, organized as a 3D array where the dimensions are [time, locations, variables].}

\item{wt}{Vector indicating the weather type for each time step.}

\item{parm}{Parameters object containing lambda transformations and other model parameters.}

\item{names}{Names of the variables included in the simulation.}
}
\value{
Simulated data with inverse transformations applied to bring it back to its original scale.
}
\description{
Function to apply inverse transformations to the simulated data, reversing the lambda transformations applied during simulation.
}
\keyword{internal}
