% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_helper_functions.R
\name{assign_seasons}
\alias{assign_seasons}
\title{Assign Seasons}
\usage{
assign_seasons(dates, seasons)
}
\arguments{
\item{dates}{A vector of dates to classify into seasons.}

\item{seasons}{A list where each element represents a season with its start and end dates and months.}
}
\value{
A numeric vector where each element represents the season assigned to the corresponding date in the input vector.
}
\description{
Assigns a season to each date based on predefined season boundaries.
}
\keyword{internal}
