% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_helper_functions.R
\name{calculate_Bk_matrices}
\alias{calculate_Bk_matrices}
\title{Calculate Bk Matrices for Autoregressive (AR) Model}
\usage{
calculate_Bk_matrices(C_k_matrices, Bk_0)
}
\arguments{
\item{C_k_matrices}{A list of covariance matrices, where each matrix represents the covariance at a different time lag. 
C_k_matrices[[1]] is the covariance matrix at lag 0.}
}
\value{
A list of Bk matrices (coefficient matrices) for the AR model, including Bk_0, which is derived separately from the rest of the Bk matrices.
}
\description{
Computes the coefficient matrices (Bk) for an autoregressive (AR) model using lagged covariance matrices.
}
\keyword{internal}
