% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_helper_functions.R
\name{calculate_ICI}
\alias{calculate_ICI}
\title{Calculate Insolation Clearness Index (ICI)}
\usage{
calculate_ICI(radiation, time)
}
\arguments{
\item{radiation}{Measured radiation values.}

\item{time}{Time corresponding to the radiation measurements.}
}
\value{
The Insolation Clearness Index (ICI).
}
\description{
Computes the Insolation Clearness Index (ICI) based on the provided radiation and time.
}
\keyword{internal}
