% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{check_parameters_validity}
\alias{check_parameters_validity}
\title{Check parameters validities}
\usage{
check_parameters_validity(parameters_values, parameters_id)
}
\arguments{
\item{parameters_values}{a vector of parameters ordered as in parameters.R}

\item{parameters_id}{a vector of parameters to check}
}
\value{
TRUE if all parameters are in theirs boundatries, otherwise FALSE
}
\description{
Check specific parameters `parameters_id` in a vector of parameters values 
`parameters_values` ordered as in parameters.R indices.
}
