% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{compute_rho2}
\alias{compute_rho2}
\title{Compute rho2 Correlations}
\usage{
compute_rho2(parm, names, cr)
}
\arguments{
\item{parm}{A data frame or list containing parameters for the Gneiting function.}

\item{names}{A vector of variable names (e.g., "temperature", "wind") to calculate correlations between.}

\item{cr}{Matrix of initial correlation values between the variables.}
}
\value{
Symmetric matrix of adjusted correlation coefficients (beta) between the variables.
where each element [i, j] represents the correlation coefficient
 between variables i and j, adjusted based on the Gneiting function and a correction term.
}
\description{
This function calculates the beta correlation coefficients between variables based on the Gneiting function, adjusted for a correction term. It is intended for internal use within package functions to adjust initial correlation values using specified parameters.
}
\details{
This function implements the methods described in Section 2.4 in Equation 8 of the article
\strong{Stochastic Environmental Research and Risk Assessment, 2025} (DOI: 10.1007/s00477-024-02897-8).
}
\keyword{internal}
