% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{cov_matrices}
\alias{cov_matrices}
\title{Generate Covariance Matrices for Spatio-Temporal Model}
\usage{
cov_matrices(par, coordinates, names, M)
}
\arguments{
\item{par}{Parameters for the Gneiting covariance function, including details for variable pairs.}

\item{coordinates}{Matrix or data frame containing spatial coordinates for each location.}

\item{names}{Vector of variable names involved in the covariance calculations.}

\item{M}{Maximum time lag considered in the model.}
}
\value{
A list of covariance matrices for each time lag up to M, and for each pair of variables, where each matrix represents the spatial covariance structure for a given time lag and variable pair.
}
\description{
Creates covariance matrices for each time lag and pair of variables using Gneiting's function, based on provided model parameters and spatial locations. These matrices are essential for multivariate space-time modeling, reflecting the covariance structure across space and time.
}
\keyword{internal}
