% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data}
\alias{data}
\title{SAFRAN meteorological data for a region in southern France for precipitation, wind and maximum temperature.}
\format{
a 3-dimensional array [days,coordinates,variables] as index
\describe{
  \item{days index}{first dimension : 4018 days index from 2011/01/01 to 2021/12/31}
  \item{coordinates index}{second dimension : 108 coordinates index}
  \item{variables index}{third dimension : 3 variables index precipitation, wind and maximum temperature}
}
}
\description{
SAFRAN meteorological data for a region in southern France for precipitation, wind and maximum temperature.
}
\examples{
data("data", package="MSTWeatherGen")
data("coordinates", package = "MSTWeatherGen")
variables = c("Precipitation", "Wind", "Temp_max")
days = seq(as.Date("2011-01-01"),as.Date("2021-12-31"), by="day")
str(data)
}
\references{
\url{https://www.umr-cnrm.fr/spip.php?article788}
}
\author{
Said Obakrim
}
\keyword{data}
