% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{dates}
\alias{dates}
\title{Sequence of dates from the data}
\format{
a vector with 4018 dates from 2011-01-01 to 2021-12-31
}
\description{
Sequence of dates from the data
}
\examples{
data("dates", package = "MSTWeatherGen")
str(dates)
}
\author{
Said Obakrim
}
