% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTWeatherGen-package.R
\name{.onAttach}
\alias{.onAttach}
\title{Things to do at package attach}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{a character string giving the library directory where
the package defining the namespace was found.}

\item{pkgname}{a character string giving the name of the package.}
}
\description{
Print package information and check dependencies
}
