% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRF_helper_functions.R
\name{estimate_gaussian_field_params}
\alias{estimate_gaussian_field_params}
\title{Estimation of Gaussian Field Parameters}
\usage{
estimate_gaussian_field_params(
  data,
  wt,
  names,
  coordinates,
  tmax,
  max_it,
  n1,
  n2,
  dates,
  threshold_precip
)
}
\arguments{
\item{data}{A 3D array containing weather data with dimensions [time, location, variable].}

\item{wt}{Vector of weather type classifications for each time point in the data.}

\item{names}{Vector of variable names in the data array.}

\item{coordinates}{A matrix of geographic coordinates for the locations in the data.}

\item{tmax}{Maximum time lag for temporal analysis.}

\item{max_it}{Maximum number of iterations for optimization procedures.}

\item{n1, }{n2 Parameters defining the granularity of spatial analysis.}

\item{dates}{Vector of dates corresponding to the time points in the data.}

\item{threshold_precip}{Threshold values for considering precipitation, used in data preprocessing.}
}
\value{
A list of estimated Gaussian field parameters for each weather type.
}
\description{
Estimates the parameters of a Gaussian field model for each weather type across spatial and temporal dimensions of weather data.
}
\keyword{internal}
