% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{estimate_lambda_transformations}
\alias{estimate_lambda_transformations}
\title{Estimate Lambda Transformations}
\usage{
estimate_lambda_transformations(data, wt, names, coordinates)
}
\arguments{
\item{data}{Input weather data, organized as a 3D array where the dimensions are [time, locations, variables].}

\item{wt}{Vector indicating the weather type for each observation.}

\item{names}{Names of the variables in the data.}

\item{coordinates}{Spatial coordinates corresponding to each location in the data.}
}
\value{
A list containing lambda transformations for each variable, location, and weather type, along with the thresholds for precipitation.
}
\description{
Function to estimate lambda transformations for each variable, location, and weather type based on the data.
}
\keyword{internal}
