% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather_types.R
\name{estimate_transitions}
\alias{estimate_transitions}
\title{Estimate Transition Matrices}
\usage{
estimate_transitions(cluster, dates, nb = 30, K)
}
\arguments{
\item{cluster}{A vector of cluster assignments for each time point.}

\item{dates}{A vector of dates corresponding to each time point.}

\item{nb}{Base number of neighbors to consider for each year, default is 30.}

\item{K}{The number of clusters, which determines the size of the transition matrices.}
}
\value{
A list of transition matrices, each corresponding to a time point. These matrices contain
  the estimated probabilities of transitioning from one cluster to another based on the nearest neighbors' approach.
}
\description{
Function to estimate transition matrices between clusters over time.
}
\keyword{internal}
