% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather_types.R
\name{estimtransition}
\alias{estimtransition}
\title{Transition Matrix Estimation}
\usage{
estimtransition(wt, dates, K)
}
\arguments{
\item{wt}{A vector indicating the weather type.}

\item{dates}{A vector of dates corresponding to each weather type.}

\item{K}{The total number of unique weather types.}
}
\value{
A square matrix of size K x K where each element [i, j] represents the normalized probability
  of transitioning from weather type i to weather type j.
}
\description{
Function to estimate transition matrices based on weather types.
}
\keyword{internal}
