% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{extract_rho1}
\alias{extract_rho1}
\title{Extract Correction Terms Matrix}
\usage{
extract_rho1(parm, names)
}
\arguments{
\item{parm}{A data frame or list containing the model parameters, including 'rho1ij' values.}

\item{names}{Character vector specifying the variable names for which correction terms are to be calculated.}
}
\value{
A square matrix where each element [i, j] represents the correction term ('rho1ij') between the ith and jth variables, facilitating the adjustment of correlations or covariances between them.
}
\description{
Extracts a matrix of correction terms ('rho1ij') for each pair of variables based on the model parameters provided in 'parm'. Designed for internal use to facilitate calculations involving correction terms in spatial or spatio-temporal modeling.
}
\details{
This function implements the methods described in Sections 2.4 in Equation 8 of the article
\strong{Stochastic Environmental Research and Risk Assessment, 2025} (DOI: 10.1007/s00477-024-02897-8).
}
\keyword{internal}
