% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_helper_functions.R
\name{filter_season_data}
\alias{filter_season_data}
\title{Filter Seasonal Data}
\usage{
filter_season_data(data, dates, season, names)
}
\arguments{
\item{data}{A 3D array of weather data with dimensions [time, location, variable].}

\item{dates}{A vector of dates corresponding to the time dimension of the data.}

\item{season}{A list specifying the season with 'min_day', 'max_day', 'min_month', and 'max_month'.}

\item{names}{Names of the variables in the data array to be used for analysis.}
}
\value{
A list containing the filtered weather data ('data_filtered') and corresponding dates ('dates_filtered').
}
\description{
Filters weather data and corresponding dates for a specified season.
}
\keyword{internal}
