% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_helper_functions.R
\name{generate_initial_conditions}
\alias{generate_initial_conditions}
\title{Generate Initial Conditions for AR Process}
\usage{
generate_initial_conditions(AR_lag, bk, wt)
}
\arguments{
\item{AR_lag}{The lag order of the AR model.}

\item{bk}{A list of Bk matrices for the AR model, one for each weather type.}

\item{wt}{The initial weather type (state) for the simulation.}
}
\value{
A matrix representing the initial conditions for the AR process simulation.
}
\description{
Generates initial conditions for the autoregressive (AR) process based on Bk matrices and initial weather type.
}
\keyword{internal}
