% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRF_helper_functions.R
\name{generate_spatial_index_pairs}
\alias{generate_spatial_index_pairs}
\title{Generate Spatial Index Pairs}
\usage{
generate_spatial_index_pairs(coordinates, n1, n2)
}
\arguments{
\item{coordinates}{A matrix or data frame containing the coordinates of points.}

\item{n1}{The number of points to uniformly distribute across the spatial domain.}

\item{n2}{The number of points to select based on proximity to each of the n1 points.}
}
\value{
A matrix where each row represents a pair of indices corresponding to the spatial location pairs selected for analysis.
}
\description{
Generates pairs of spatial indices based on a subset of uniformly distributed points and additional points selected based on proximity.
This function first selects a set of n1 points uniformly distributed across the spatial domain. For each of these points,
it then selects n2 points based on their proximity, ensuring a diverse yet focused selection of spatial pairs for further analysis.
}
\keyword{internal}
