% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRF_helper_functions.R
\name{generate_temporal_index_pairs}
\alias{generate_temporal_index_pairs}
\title{Generate Temporal Index Pairs}
\usage{
generate_temporal_index_pairs(wt_id, dates, tmax)
}
\arguments{
\item{wt_id}{Identifiers (indices) corresponding to specific time points in the dataset.}

\item{dates}{Vector of dates corresponding to the time dimension in the dataset.}

\item{tmax}{Maximum temporal lag for which pairs are to be generated.}
}
\value{
A matrix with each row representing a pair of temporal indices and their corresponding time lag. The columns 't1' and 't2'
represent the indices of the paired time points, and 'u' represents the time lag between them.
}
\description{
Creates pairs of temporal indices based on specified time lags up to a maximum time lag (tmax). This function is useful for
analyzing temporal relationships and covariances at different lags in spatio-temporal data.
}
\keyword{internal}
