% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRF_helper_functions.R
\name{generate_variable_index_pairs}
\alias{generate_variable_index_pairs}
\title{Generate Variable Index Pairs}
\usage{
generate_variable_index_pairs(names)
}
\arguments{
\item{names}{A vector of variable names for which index pairs are to be generated.}
}
\value{
A data frame with two columns ('v1' and 'v2') listing all unique pairs of variables. The data frame includes both
pairs of different variables and pairs of the same variable.
}
\description{
Generates a data frame containing all unique pairs of variable names, including pairs of the same variable. This is used for
generating covariance matrices and parameter matrices where interactions between different variables (or the same variable) are
considered.
}
\keyword{internal}
