% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTWeatherGen-package.R
\name{getCores}
\alias{getCores}
\title{Get the information of number of Cores for Parallel Processing for the user.}
\usage{
getCores()
}
\value{
Integer. The number of cores currently configured for parallel processing.
}
\description{
Retrieves the current number of cores configured for parallel computations.
If no cores have been set yet, it automatically initializes them by calling 
setCores().
}
\seealso{
setCores() to configure the number of cores.
}
