% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTWeatherGen-package.R
\name{getInfo}
\alias{getInfo}
\title{Get Package and System Information}
\usage{
getInfo()
}
\description{
Displays information about the MSTWeatherGen package and system configuration.
This includes the package name, version, license, and the number of CPU cores 
available on your computer.
}
\seealso{
[setCores()], [getCores()] for configuring parallel processing cores.
}
