% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{Gneiting}
\alias{Gneiting}
\title{Gneiting's Spatio-Temporal Covariance Model}
\usage{
Gneiting(h, u, par, dij)
}
\arguments{
\item{h}{Numeric vector of spatial distances.}

\item{u}{Numeric vector of temporal distances.}

\item{par}{Numeric vector of parameters for the covariance function.}

\item{dij}{Correlation parameter between variables i and j.}
}
\value{
Numeric vector of covariance values calculated using Gneiting's model.
}
\description{
Computes the covariance based on Gneiting's spatio-temporal model. Intended for internal package use.
}
\details{
This function implements the methods described in Section 2.4 in Equations 8 of the article
*Stochastic Environmental Research and Risk Assessment, 2025* (DOI: 10.1007/s00477-024-02897-8).
}
\keyword{internal}
