% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_helper_functions.R
\name{haversine}
\alias{haversine}
\title{Haversine Distance Calculation}
\usage{
haversine(point1, point2)
}
\arguments{
\item{point1}{Coordinates of the first point in decimal degrees (lon, lat).}

\item{point2}{Coordinates of the second point in decimal degrees (lon, lat).}
}
\value{
The Haversine distance between the two points in kilometers.
}
\description{
Computes the Haversine distance between two points given their coordinates.
}
\keyword{internal}
