% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRF_helper_functions.R
\name{init_space_par}
\alias{init_space_par}
\title{Initialize Spatial Parameters for Variables}
\usage{
init_space_par(data, names, h, uh, max_it = 2000)
}
\arguments{
\item{data}{The dataset for which spatial parameters are being initialized, typically
a 3D array or a list of spatial observations for multiple variables.}

\item{names}{A vector of variable names for which spatial parameters are to be initialized.}

\item{h}{Vector of spatial distances used in the likelihood function, representing
the spatial relationship between observation points.}

\item{uh}{Matrix containing additional data required by the likelihood function, 
typically combining spatial and temporal distances with other relevant information.}

\item{max_it}{Maximum number of iterations for the optimization process, defaulting to 2000.
This parameter controls the depth of the optimization search.}
}
\value{
A list containing optimized spatial parameters for each variable specified in 'names'.
        Each element of the list corresponds to a set of parameters (e.g., range and smoothness)
        optimized for the spatial structure of the corresponding variable.
}
\description{
Optimizes initial spatial parameters for each variable in a dataset. This function employs
a log-likelihood based optimization approach to determine initial values of spatial parameters
that best fit the spatial structure of the data for each variable.
}
\keyword{internal}
