% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{loglik_pair}
\alias{loglik_pair}
\title{Compute Log-likelihood for Variable Pair}
\usage{
loglik_pair(par, parms, pair, par_all, data, names, Vi, h, u, uh, ep, cr)
}
\arguments{
\item{par}{Current parameters being optimized.}

\item{parms}{Indices or names of parameters in 'par' to be updated.}

\item{pair}{A string indicating the pair of variables (e.g., "temperature-wind") being analyzed.}

\item{par_all}{Complete set of parameters for the model.}

\item{data}{3D array of observed data, with dimensions corresponding to times, locations, and variables.}

\item{names}{Vector of variable names, indicating the variables' names (e.g., "temperature" and "wind").}

\item{Vi}{Matrix where each line corresponds for a possible combination of variables in "names".}

\item{h}{Vector of spatial distances for the pair.}

\item{u}{Vector of temporal distances for the pair.}

\item{uh}{Matrix containing pairs of spatial and temporal distances, and additional information.}

\item{ep}{A matrix or data frame defining pairs of variables.}

\item{cr}{Correlation matrix, initial or base correlations between variables.}
}
\value{
The log-likelihood value for the given pair of variables based on the current model parameters.
}
\description{
Calculates the log-likelihood for a given pair of variables using the Gneiting spatio-temporal covariance model. This function is part of the internal mechanism for optimizing model parameters based on observed data.
}
\keyword{internal}
