% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{Matern}
\alias{Matern}
\title{Matern Covariance Function}
\usage{
Matern(h, r, v)
}
\arguments{
\item{h}{Numeric vector of distances between points.}

\item{r}{Scalar range parameter of the Matern function, affecting the spatial correlation decay.}

\item{v}{Scalar smoothness parameter of the Matern function, controlling the smoothness of the resulting field.}
}
\value{
Numeric vector representing the covariance values calculated using the Matern function for the distances in `h`.
}
\description{
Calculates the Matern covariance function for a given vector of distances. This function is intended for internal use within the package to compute spatial covariances with the Matern model. It is not exported for end-user interaction.
}
\keyword{internal}
