% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{modify_beta_parm}
\alias{modify_beta_parm}
\title{Modify Beta Parameters in Model Parameters}
\usage{
modify_beta_parm(parm, beta)
}
\arguments{
\item{parm}{A data frame or list representing the current set of model parameters, including 'v1', 'v2', and 'dij' among others.}

\item{beta}{A matrix of beta coefficients computed to adjust the correlations or dependencies between variables.}
}
\value{
The modified set of model parameters with updated 'dij' values based on the beta coefficients.
}
\description{
Adjusts the 'dij' parameters in the model parameter set based on computed beta coefficients, ensuring that the covariance structure reflects these adjustments.
}
\keyword{internal}
