% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_helper_functions.R
\name{most_probable_weather_type}
\alias{most_probable_weather_type}
\title{Predict Most Probable Weather Type}
\usage{
most_probable_weather_type(sim, centroids, transitions, names_weather_types)
}
\arguments{
\item{sim}{A matrix containing simulated weather data for a single time step.}

\item{centroids}{A list of centroids, where each centroid is a vector representing the mean values of weather variables for a specific weather type.}

\item{transitions}{A list (or matrix) of transition probabilities between weather types. Each element transitions[[i]][j, k] represents the probability of transitioning from weather type j to k.}

\item{names_weather_types}{Names of the weather variables used to determine the weather type.}
}
\value{
The predicted weather type for the next time step, selected based on the calculated probabilities.
}
\description{
Predicts the most likely weather type for the next time step in a simulated weather series.
}
\keyword{internal}
