% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTWeatherGen-package.R
\docType{package}
\encoding{UTF-8}
\name{MSTWeatherGen-package}
\alias{MSTWeatherGen-package}
\alias{MSTWeatherGen}
\title{Multivariate Space-Time Weather Generator}
\description{
Simulation model of multivariate spatio-temporal meteorological variables.
It provides tools for estimating the model parameters and generating
}
\details{
\tabular{ll}{
         Package: \tab MSTWeatherGen\cr
         Type: \tab Package\cr
         Version: \tab 1.0.0\cr
         Date: \tab 2026-18-02\cr
         License: \tab GPL (>=3)\cr
         }
}
\examples{
\dontrun{
library("MSTWeather")

}
}
\references{
## When referencing the simulation model, please cite the following article:


## When referencing the R package, please cite the following package:

url: https://cran.r-project.org/package=MSTWeatherGen.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Chair-Geolearning/MSTWeatherGen}
  \item \url{https://chair-Geolearning.github.io/MSTWeatherGen}
  \item Report bugs at \url{https://github.com/Chair-Geolearning/MSTWeatherGen/issues}
}

}
\author{
Said Obakrim \email{saidobak@gmail.com}

Jean-François Rey \email{jean-francois@inrae.fr}

Ahmed Boualam \email{ahmed.boualam@inrae.fr}

Denis Allard \email{denis.allard@inrae.fr}

Lionel Benoit \email{lionel.benoit@inrae.fr}

Maintainer: Jean-FRançois Rey \email{jean-francois.rey@inrae.fr}
}
\keyword{internal}
