% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTWeatherGen.R
\name{MSTWeatherGen_Estim_season}
\alias{MSTWeatherGen_Estim_season}
\title{Seasonal Estimation for MSTWeatherGen}
\usage{
MSTWeatherGen_Estim_season(
  data,
  dates,
  precipitation = T,
  scale = FALSE,
  names = NULL,
  names_weather_types = NULL,
  coordinates,
  season,
  max_it,
  tmax,
  n1,
  n2
)
}
\arguments{
\item{data}{Array of weather data with dimensions [time, location, variable].}

\item{dates}{Vector of dates corresponding to the time dimension of the data.}

\item{precipitation}{Logical indicating if precipitation is to be considered as a primary variable.}

\item{scale}{Logical, indicating if the data needs to be standardized (`TRUE`) or not (`FALSE`). If `scale` is `TRUE` each meteorological variable
is standardized at each location using smoothed mean and standard deviation.}

\item{names}{Names of the variables in the data array to be used for analysis.}

\item{names_weather_types}{Subset of 'names', variables to be used for weather type classification.}

\item{coordinates}{Matrix with columns for the coordinates of each location.}

\item{season}{Vector of integers representing months to define the season for analysis.}

\item{max_it}{Maximum number of iterations for optimization procedures.}

\item{tmax}{Maximum time lag for temporal analysis.}

\item{n1, }{n2 Parameters defining spatial window size for analysis.}
}
\value{
A list containing dates, scale parameters, weather types, transition probabilities, lambda transformations,
and parameters for the Gaussian field model for the specified season.
}
\description{
Performs seasonal estimation for the MSTWeatherGen package, which includes identification of weather types,
estimation of scaling parameters, computation of transition probabilities between weather types, and estimation
of parameters for the Gaussian field model.
}
\keyword{internal}
