% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRF_helper_functions.R
\name{optimize_spatial_parameters}
\alias{optimize_spatial_parameters}
\title{Optimize Spatial Parameters for Variable Pairs}
\usage{
optimize_spatial_parameters(par_all, data, names, Vi, uh, cr, max_it)
}
\arguments{
\item{par_all}{Initial set of all model parameters before optimization.}

\item{data}{The dataset used for optimization, typically a 3D array or a list
of spatial observations for multiple variables over time.}

\item{names}{Vector of variable names involved in the optimization process.}

\item{Vi}{Matrix indicating the pairs of variables for which parameters are optimized.}

\item{uh}{Matrix containing additional data required by the likelihood function,
combining spatial and temporal distances with other relevant information.}

\item{cr}{Initial correlation matrix used in the optimization process to provide
a starting point for parameter estimation.}

\item{max_it}{Maximum number of iterations allowed for the optimization algorithm,
controlling the depth of the search process.}

\item{ep}{Data frame defining pairs of variables for analysis and optimization,
used to guide the optimization process by specifying which variable interactions to consider.}
}
\value{
A vector `par_all` updated with optimized model parameters for each variable pair.
        This vector consolidates the model parameters post-optimization, ready for model application or further analysis.
}
\description{
Performs optimization of model parameters for each pair of variables based on spatial data.
This function iteratively optimizes spatial parameters to maximize the log-likelihood
of observing the data given the model, focusing on the spatial interaction between pairs of variables.
}
\details{
This function implements the methods described in Section 3.3 of the article, functions log_lik and log
*Stochastic Environmental Research and Risk Assessment, 2025* (DOI: 10.1007/s00477-024-02897-8).
}
\keyword{internal}
