% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{orderNorm}
\alias{orderNorm}
\title{Ordered Normalization}
\usage{
orderNorm(x, left, n_logit_fit = min(length(x), 1e+05), ..., warn = TRUE)
}
\arguments{
\item{x}{Numeric vector to be normalized.}

\item{left}{Numeric value specifying the left truncation.}

\item{n_logit_fit}{The number of points to use in the logistic regression for extrapolation, default is the lesser of the length of x or 100,000.}

\item{...}{Additional arguments passed to the glm function.}

\item{warn}{Logical flag indicating whether to issue a warning when ties are detected in the data.}
}
\value{
A list with class 'orderNorm' containing the normalized data (x.t), the original data (x), and a fitted model object (fit) for future extrapolation.
}
\description{
Function for ordered normalization using Ordered Quantile Normalizing transformation.
}
\keyword{internal}
