% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{orderNormTransf}
\alias{orderNormTransf}
\title{Transforms New Data Points using orderNorm Object}
\usage{
orderNormTransf(orderNorm_obj, new_points, warn, left)
}
\arguments{
\item{orderNorm_obj}{Fitted orderNorm object containing details of the original normalization or transformation.}

\item{new_points}{New data points to be transformed.}

\item{warn}{Logical indicating whether to issue warnings for transformations outside the observed domain.}

\item{left}{Parameter affecting the normalization method used in the original transformation.}
}
\value{
Transformed data points based on the fitted model.
}
\description{
Function to transform new data points based on a previously fitted normalization or transformation model.
}
\keyword{internal}
