% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{pd_condition}
\alias{pd_condition}
\title{Check Positive Definiteness Condition}
\usage{
pd_condition(parm, names)
}
\arguments{
\item{parm}{A data frame or list containing the model parameters.}

\item{names}{Character vector of variable names for which the condition is checked.}
}
\value{
Logical value indicating whether the covariance matrix, constructed based on the parameters and variable names, is positive definite.
}
\description{
Verifies the positive definiteness of the covariance matrix constructed from model parameters, which is crucial for ensuring valid covariance structures in spatio-temporal modeling.
}
\keyword{internal}
