% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_plots.R
\name{plot_dry_wet_spells_maps}
\alias{plot_dry_wet_spells_maps}
\title{Plot Dry and Wet Spells Maps}
\usage{
plot_dry_wet_spells_maps(sim, observed, coordinates, dates)
}
\arguments{
\item{sim}{A 3-dimensional array of simulated weather data with dimensions [time, location, variable], where "Precipitation" is expected to be one of the variables.}

\item{observed}{A 3-dimensional array of observed weather data with the same dimensions and variable expectations as `sim`.}

\item{coordinates}{A matrix or data frame of geographic coordinates for the locations represented in `sim` and `observed`, with longitude in the first column and latitude in the second.}

\item{dates}{A vector of dates corresponding to the time dimension in `sim` and `observed`. This parameter is currently not used in the function but may be intended for future use or filtering within the provided data.}
}
\value{
A `ggplot` object representing the maps of dry and wet spells for observed and simulated precipitation data across the specified locations.
}
\description{
Generates maps showing the frequency of consecutive wet days for observed and simulated precipitation data. This function uses precipitation data to classify days as wet or dry based on a threshold and calculates the length of consecutive wet day spells. The results are then plotted on a map, with points colored by the number of spells and faceted by observation type and spell length category.
}
\examples{
\dontrun{
  # Assuming `sim`, `observed`, and `coordinates` are already defined and correctly formatted:
  plot_dry_wet_spells_maps(sim, observed, coordinates, dates)
}
}
