% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_plots.R
\name{plot_mean_by_month}
\alias{plot_mean_by_month}
\title{Plot Monthly Mean Comparisons}
\usage{
plot_mean_by_month(sim, observed, places, names_places, names, dates)
}
\arguments{
\item{sim}{A 3-dimensional array of simulated weather data with dimensions [time, location, variable].}

\item{observed}{A 3-dimensional array of observed weather data with the same dimensions as `sim`.}

\item{places}{A vector of indices representing the locations for which the data will be plotted. If `places` has only one element, the function generates plots for that specific location; otherwise, it generates plots for each location specified.}

\item{names_places}{A vector of names corresponding to the `places` indices, used for labeling in the plots.}

\item{names}{A vector of variable names to be included in the comparison. These should correspond to the third dimension of the `sim` and `observed` arrays.}

\item{dates}{A vector of dates corresponding to the first dimension of the `sim` and `observed` arrays. This is used to determine the months and years for grouping the data.}
}
\value{
A `ggplot` object containing the boxplots comparing the monthly mean values of the observed and simulated data for the specified variables and locations.
}
\description{
Compares the monthly mean values of observed and simulated weather data for specified variables and locations. The function generates boxplots to visually compare the means for each month, facilitating an easy comparison between observed and simulated data across different variables and locations.
}
\examples{
\dontrun{
  # Assuming `sim`, `observed`, `dates`, `places`, and `names` are predefined:
  plot <- plot_mean_by_month(sim, observed, places, names_places, names, dates)
  print(plot)
}

}
