% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_plots.R
\name{plot_multivariate_st_covariances}
\alias{plot_multivariate_st_covariances}
\title{Plot multivariate space-time covariances}
\usage{
plot_multivariate_st_covariances(vgm, names)
}
\description{
Function plot multivariate space-time covariances.
}
\keyword{internal}
