% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{predict.orderNormTransf}
\alias{predict.orderNormTransf}
\title{Predict Method for orderNormTransf Objects}
\usage{
\method{predict}{orderNormTransf}(object, newdata = NULL, inverse = FALSE, warn = TRUE, ...)
}
\arguments{
\item{object}{Fitted orderNormTransf object.}

\item{newdata}{Numeric vector of new data to be transformed. If NULL, transformation is applied to the original data.}

\item{inverse}{Logical indicating whether to perform inverse transformation. Default is FALSE.}

\item{warn}{Logical indicating whether to issue a warning when ties are detected in the data. Default is TRUE.}

\item{...}{Additional arguments to be passed.}
}
\value{
Transformed or inverse-transformed data based on the specified operation.
}
\description{
Function to predict the normalized values for new data using a fitted orderNormTransf object or 
to perform inverse transformation.
}
\keyword{internal}
