% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRF_helper_functions.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Data Preprocessing for Spatio-Temporal Models}
\usage{
preprocess_data(Ti, Si, coordinates)
}
\arguments{
\item{Ti}{A matrix containing temporal index pairs and their respective lags.}

\item{Si}{A matrix containing spatial index pairs.}

\item{coordinates}{A matrix of spatial coordinates for each location.}
}
\value{
A list containing the preprocessed data, including a combined matrix of temporal lags, spatial distances, and indices (`uh`), as well as separate vectors of temporal lags (`u`) and spatial distances (`h`).
}
\description{
Performs preprocessing on spatio-temporal data to prepare for model fitting, including computing spatial distances and temporal lags.
}
\keyword{internal}
