% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{scale_data}
\alias{scale_data}
\title{Scale Data}
\usage{
scale_data(data, names, dates, window_size = 30)
}
\arguments{
\item{data}{Input weather data, organized as a 3D array where the dimensions are [time, locations, variables].}

\item{names}{Names of the variables in the data.}

\item{dates}{Vector of dates corresponding to the time dimension in the data.}

\item{window_size}{Size of the window for the moving average used to smooth the daily mean and standard deviation. Default is 30.}
}
\value{
A list containing the scaled data and scale parameters (mean and standard deviation) for each variable.
}
\description{
Function to scale the input data based on daily mean and standard deviation, and optionally smooth them using a moving average.
}
\keyword{internal}
