% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRF_helper_functions.R
\name{selectPoints}
\alias{selectPoints}
\title{Select Points Based on Probabilities}
\usage{
selectPoints(coordinates, betaIndex, v)
}
\arguments{
\item{coordinates}{A matrix or data frame containing the coordinates of points.}

\item{betaIndex}{The index of the reference point in the coordinates matrix/data frame.}

\item{v}{The number of points to select, including the reference point.}
}
\value{
A vector of selected indices from the coordinates matrix/data frame, including the index of the reference point.
}
\description{
Selects a set of points from a dataset based on probabilities calculated from their spatial distances to a specific reference point.
The function calculates the probability for each point as inversely proportional to its squared distance from the reference point, ensuring a higher chance of selecting closer points.
}
\keyword{internal}
