% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRF_helper_functions.R
\name{selectUniformPointsIndices}
\alias{selectUniformPointsIndices}
\title{Select Uniformly Distributed Points Indices}
\usage{
selectUniformPointsIndices(coordinates, N)
}
\arguments{
\item{coordinates}{A matrix or data frame of geographical coordinates (longitude and latitude).}

\item{N}{The number of uniformly distributed points to select.}
}
\value{
A vector of indices corresponding to the selected uniformly distributed points.
}
\description{
This function selects indices of points from a given set of coordinates that are uniformly distributed across the spatial domain.
The selection process involves performing k-means clustering on the coordinates and then selecting the closest data point to each cluster center.
}
\keyword{internal}
