% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTWeatherGen-package.R
\name{setCores}
\alias{setCores}
\title{Set Number of Cores for Parallel Processing}
\usage{
setCores(n = NULL)
}
\arguments{
\item{n}{The number of cores to use. If  NULL (default), 
automatically sets to total_cores - 2 (minimum 1). If specified, 
must be at least 1 and will be capped by the maximum which will be at the total number of 
available cores.}
}
\value{
Number of cores set (invisible)
}
\description{
Function to set automatically or manually the number of cores
It sets automatically the numbers of core as follows : nb_of_cores available - 2
To change manually you need to call it and put the
}
\examples{
#Automatic setting setCores()
#Manual setting to 4 cores : setCores(4)
}
