% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{spacetime_cov}
\alias{spacetime_cov}
\title{Compute Spatio-Temporal Covariances}
\usage{
spacetime_cov(
  data,
  wt_id,
  locations,
  ds = NULL,
  dates,
  lagstime,
  dist,
  covgm = TRUE
)
}
\arguments{
\item{data}{3D array representing time, location, and variable dimensions of the spatio-temporal data.}

\item{wt_id}{Indices of weather types for which covariances are computed.}

\item{locations}{Matrix of spatial locations for the data points.}

\item{ds}{Precomputed distance matrix or NULL to compute distances from 'locations'.}

\item{dates}{Vector of dates corresponding to the time dimension of the data.}

\item{lagstime}{Vector of time lags for covariance computation.}

\item{dist}{Vector of spatial distances for covariance computation.}

\item{covgm}{Logical flag to compute cross-covariances (default TRUE).}
}
\value{
Data frame containing computed covariances for specified spatial distances and time lags, facilitating the analysis of spatial and temporal patterns in the data.
}
\description{
Calculates spatial and temporal covariances for given spatio-temporal data, facilitating the understanding of spatial and temporal variability in the context of different weather types.
}
\keyword{internal}
