% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_helper_functions.R
\name{summary.swg}
\alias{summary.swg}
\title{Explore or summarize estimation output swg}
\usage{
\method{summary}{swg}(swg, season = NULL, all = FALSE)
}
\arguments{
\item{swg}{A swg class, as a list containing the results of the
 `MSTWeatherGen_Estim` function, including estimated parameters and
 other outputs relevant to weather generation, such as weather type
 classifications and spatial dependencies.
Each season contains a list containing dates, scale parameters,
 weather types, transition probabilities, lambda transformations,
and parameters for the Gaussian field model for the specified season.}

\item{season}{vector of seasons names,
default NULL show summarize seasons}

\item{all}{a boolean to output all values as str (default = FALSE)}
}
\value{
invisible(NULL)
}
\description{
Summarize `MSTWeatherGen_Estim()`` function output (swg object class).
}
\examples{
\dontrun{
  swg <- MSTWeatherGen_estim()
  summary(swg)
  summary.swg(swg)
}
}
