% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{transformations}
\alias{transformations}
\title{Transform Data using Lambda Transformations}
\usage{
transformations(data, wt, names, coordinates, lmbd)
}
\arguments{
\item{data}{Input weather data, organized as a 3D array where the dimensions are [time, locations, variables].}

\item{wt}{Vector indicating the weather type for each observation.}

\item{names}{Names of the variables in the data.}

\item{coordinates}{Spatial coordinates corresponding to each location in the data.}

\item{lmbd}{Lambda transformation parameters estimated for each variable, location, and weather type.}
}
\value{
Transformed data using lambda transformations.
}
\description{
Function to transform data using lambda transformations based on the provided lambda parameters.
}
\keyword{internal}
