% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRF_helper_functions.R
\name{unique_elements}
\alias{unique_elements}
\title{Unique Elements Extraction}
\usage{
unique_elements(Si)
}
\arguments{
\item{Si}{A vector or list from which unique elements need to be extracted.}
}
\value{
A vector containing only unique elements from the input.
}
\description{
Extracts unique elements from a vector or list, effectively removing duplicates.
}
\details{
This function implements the methods described in Section 3.3 of the article
*Stochastic Environmental Research and Risk Assessment, 2025* (DOI: 10.1007/s00477-024-02897-8).
}
\keyword{internal}
