% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{update_rho2_parameters}
\alias{update_rho2_parameters}
\title{update rho2ij values in a names vector}
\usage{
update_rho2_parameters(par_all, names, rho2ij)
}
\arguments{
\item{par_all}{a named vector containing variables pairs rho2ij values and models values.}

\item{names}{character vector specifying the variable names for which rho2ij is in vector rho2ij and pair have to be updates.}

\item{rho2ij}{a square matrix containing rho2ij values indiced by pairs number}
}
\value{
a named vector with variables pair rho2ij vlaues updates and the other model values.
}
\description{
extract values from rho2ij matrix for each variables pairs and save it in a named vector as 'variable1-variable2:rho2ij' index.
}
\keyword{internal}
