% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather_types.R
\name{weather_types}
\alias{weather_types}
\title{Weather Types Identification and Visualization}
\usage{
weather_types(
  data,
  variables,
  dates,
  n_wt = NULL,
  coordinates,
  max_number_wt = NULL,
  threshold = 0,
  return_plots = T,
  names_units,
  dir
)
}
\arguments{
\item{data}{An array with dimensions [time, locations, variables] containing meteorological data.}

\item{variables}{Names of variables used for constructing weather types, e.g., wind, temperature.}

\item{dates}{Vector of dates corresponding to the first dimension of the data.}

\item{n_wt}{Optional; predefined number of weather types to identify.}

\item{coordinates}{Data frame (or matrix) containing coordinates of the locations.}

\item{max_number_wt}{Optional; maximum number of weather types for automatic identification.}

\item{threshold}{Threshold used for defining wet days in precipitation analysis (default=0).}

\item{return_plots}{Logical indicating whether to return plots.}

\item{names_units}{Units of the variables for labeling plots.}

\item{dir}{Directory where to save the generated plots.}
}
\value{
A list containing the classification of each time point into weather types,
  and optionally, plots illustrating the weather types.
}
\description{
Function to identify and visualize different weather types (WTs) based on input data.
}
\keyword{internal}
