% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Z}
\alias{Z}
\title{Simulated latent Gaussian random field}
\format{
a 3-dimensional array [days,coordinates,variables] as index
\describe{
  \item{days index}{first dimension : 30 days index}
  \item{coordinates index}{second dimension : 456 coordinates index}
  \item{variables index}{third dimension : 3 variables index v1, v2, v3}
}
}
\description{
Simulated latent Gaussian random field
}
\examples{
data("Z", package = "MSTWeatherGen")
str(Z)
}
