% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{compute_ax}
\alias{compute_ax}
\title{Extract Correction Terms Matrix}
\usage{
compute_ax(parm, names)
}
\arguments{
\item{parm}{A data frame or list containing the model parameters, including 'ax' values.}

\item{names}{Character vector specifying the variable names for which correction terms are to be calculated.}
}
\value{
A square matrix where each element [i, j] represents the correction term ('ax') between the ith and jth variables, facilitating the adjustment of correlations or covariances between them.
}
\description{
Extracts a matrix of correction terms ('ax') for each pair of variables based on the model parameters. Designed for internal use to facilitate calculations involving correction terms in spatial or spatio-temporal modeling.
}
\keyword{internal}
