% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{compute_beta}
\alias{compute_beta}
\title{Compute Beta Correlations}
\usage{
compute_beta(parm, names, cr)
}
\arguments{
\item{parm}{A data frame or list containing parameters for the Gneiting function.}

\item{names}{Character vector of variable names to calculate correlations between.}

\item{cr}{Matrix of initial correlation values between the variables.}
}
\value{
Symmetric matrix of adjusted correlation coefficients (beta) between the variables.
}
\description{
This function calculates the beta correlation coefficients between variables based on the Gneiting function, adjusted for a correction term. It is intended for internal use within package functions to adjust initial correlation values using specified parameters.
}
\keyword{internal}
