% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{coordinates}
\alias{coordinates}
\title{Geographical coordinates of locations included in the meteorological data (a region in southern France).}
\format{
a data-frame with 108 rows as coordinates points and 2 columns as c("longitude","latitude") WGS84 (epsg:4326) projection
}
\description{
Geographical coordinates of locations included in the meteorological data (a region in southern France).
}
\examples{
data("coordinates", package = "MSTWeatherGen")
str(coordinates)
}
\references{
\url{https://www.umr-cnrm.fr/spip.php?article788}
}
\author{
Said Obakrim
}
\keyword{data}
