% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{coordinates}
\alias{coordinates}
\title{Geographical coordinates of locations included in the meteorological data.}
\description{
Geographical coordinates of locations included in the meteorological data.
}
\references{
\url{https://www.umr-cnrm.fr/spip.php?article788}
}
\author{
Said Obakrim
}
\keyword{data}
