% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data}
\alias{data}
\title{SAFRAN meteorological data for a region in southern France for precipitation, wind and maximum temperature.}
\description{
SAFRAN meteorological data for a region in southern France for precipitation, wind and maximum temperature.
}
\references{
\url{https://www.umr-cnrm.fr/spip.php?article788}
}
\author{
Said Obakrim
}
\keyword{data}
