% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather_types.R
\name{data_compression}
\alias{data_compression}
\title{Data Compression}
\usage{
data_compression(data)
}
\arguments{
\item{data}{The dataset to be compressed.}
}
\value{
The compressed data.
}
\description{
Function to compress data using the PTA3 method.
}
\keyword{internal}
