% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_helper_functions.R
\name{ds}
\alias{ds}
\title{Calculate Haversine Distance Between Two Coordinates}
\usage{
ds(i, j, coordinates)
}
\arguments{
\item{i}{Index of the first coordinate.}

\item{j}{Index of the second coordinate.}

\item{coordinates}{Matrix with columns for the longitude and latitude of each location.}
}
\value{
The Haversine distance between the two coordinates in kilometers.
}
\description{
Computes the Haversine distance between two coordinates using the geosphere package.
}
\keyword{internal}
