% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{extract_beta}
\alias{extract_beta}
\title{Extract Beta Coefficients Matrix}
\usage{
extract_beta(parm, names)
}
\arguments{
\item{parm}{A data frame or list containing the model parameters, which must include 'dij' values representing beta coefficients between pairs of variables.}

\item{names}{Character vector of variable names for which beta coefficients are to be extracted.}
}
\value{
A square matrix where each element [i, j] contains the beta coefficient ('dij') between the ith and jth variables. This matrix is crucial for modeling the interactions and dependencies between different variables in the model.
}
\description{
Extracts a matrix of beta coefficients ('dij') for each pair of variables from the provided model parameters. Intended for internal use, this function supports spatial and spatio-temporal modeling by organizing pairwise beta coefficients into a structured format.
}
\keyword{internal}
