% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_helper_functions.R
\name{find_centroids}
\alias{find_centroids}
\title{Find Centroids}
\usage{
find_centroids(data, dates, seasons, wt_seasons, names_weather_types)
}
\arguments{
\item{data}{A multi-dimensional array of weather data, with dimensions [time, locations, variables].}

\item{dates}{A vector of dates corresponding to the time dimension in the data array.}

\item{seasons}{A list defining the start and end dates of each season.}

\item{wt_seasons}{A list containing vectors of weather type classifications for each season.}

\item{names_weather_types}{Names of the weather variables used to determine the weather type centroids.}
}
\value{
A list of lists, where each inner list contains the centroids for each weather type within a specific season.
}
\description{
Calculates centroids of weather types for each season.
}
\keyword{internal}
