% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{inv_orderNorm_Transf}
\alias{inv_orderNorm_Transf}
\title{Inverse Transformation for orderNorm Objects}
\usage{
inv_orderNorm_Transf(orderNorm_obj, new_points_x_t, left, warn = FALSE)
}
\arguments{
\item{orderNorm_obj}{Fitted orderNorm object containing details of the original normalization or transformation.}

\item{new_points_x_t}{Transformed data points for which the original values are to be estimated.}

\item{left}{Parameter used in the original transformation, affecting the normalization method.}

\item{warn}{Logical indicating whether to issue warnings for transformations outside the observed domain. Default is FALSE.}
}
\value{
Estimated original data values corresponding to the transformed data points.
}
\description{
Function to perform inverse transformation or normalization applied by the orderNorm function.
}
\keyword{internal}
