% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{is_positive_definite}
\alias{is_positive_definite}
\title{Check if a matrix is positive}
\usage{
is_positive_definite(M, tol = -1e-10)
}
\arguments{
\item{M}{a square and symetric matrix}

\item{tol}{tolerance (default -1e-10)}
}
\value{
a boolean
}
\description{
This function checks if a symmetric matrix is positive definite
by verifying that all its eigenvalues are strictly positive
(greater than a small tolerance).
}
