% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{loglik}
\alias{loglik}
\title{Total Log-Likelihood Calculation}
\usage{
loglik(par, parms, par_all, data, names, Vi, h, u, uh, ep, cr)
}
\arguments{
\item{par}{Vector of parameter estimates currently being optimized.}

\item{parms}{Indices or names of parameters within 'par' that are subject to update.}

\item{par_all}{Comprehensive list of all model parameters, including those not currently being optimized.}

\item{data}{3D array of observed data across locations, times, and variables.}

\item{names}{Character vector of variable names.}

\item{Vi}{Matrix indicating all combinations of variables for analysis.}

\item{h}{Spatial distances vector for variable pairs.}

\item{u}{Temporal distances vector for variable pairs.}

\item{uh}{Combined matrix of spatial and temporal distances with additional identifiers.}

\item{ep}{Data frame defining variable pairs for analysis.}

\item{cr}{Initial correlation matrix across variables.}
}
\value{
Negatif total log-likelihood value for the observed data given the current model parameters.
}
\description{
Calculates the total log-likelihood for spatial or spatio-temporal data across all variable pairs.
Utilizes the Gneiting spatio-temporal covariance model to integrate log-likelihood contributions from each variable pair.
This function is core to the optimization process within model fitting.
}
\details{
This function implements the methods described in Section 3.3 of the article
\strong{Stochastic Environmental Research and Risk Assessment, 2025} (DOI: 10.1007/s00477-024-02897-8).
}
\keyword{internal}
