% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{loglik_spatial}
\alias{loglik_spatial}
\title{Log-Likelihood for Spatial Data}
\usage{
loglik_spatial(par, data, h, uh, v)
}
\arguments{
\item{par}{Vector containing parameters for the Matérn covariance function: range (`par[1]`) and smoothness (`par[2]`). Both parameters must be positive.}

\item{data}{3D array of observed spatial data.}

\item{h}{Vector of spatial distances between observations.}

\item{uh}{Matrix specifying indices for pairing spatial observations.}

\item{v}{Index of the variable within `data` for which the log-likelihood is computed.}
}
\value{
Log-likelihood value for the spatial data under the Matérn covariance model.
}
\description{
Calculates the log-likelihood for spatial data based on the Matérn covariance function. This function plays a pivotal role in estimating spatial parameters for geostatistical models.
}
\keyword{internal}
