% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{make_positive_definite}
\alias{make_positive_definite}
\title{Check make a matrix positive}
\usage{
make_positive_definite(M, epsilon = 1e-09)
}
\arguments{
\item{M}{a square and symetric matrix}

\item{epsilon}{threshold (default 1e-9)}
}
\value{
the matrix M positive
}
\description{
This function transforms a symmetric matrix into a positive definite matrix
by replacing any negative or excessively small eigenvalues with 
a specified small positive value (epsilon).
}
