% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRF_helper_functions.R
\name{optimize_spatiotemporal_parameters}
\alias{optimize_spatiotemporal_parameters}
\title{Optimize Spatio-Temporal Parameters for Variable Pairs}
\usage{
optimize_spatiotemporal_parameters(par_all, data, names, Vi, uh, cr, max_it)
}
\arguments{
\item{par_all}{Initial comprehensive set of model parameters to be refined through optimization.}

\item{data}{Dataset containing spatio-temporal observations, typically a 3D array or list
with dimensions representing locations, times, and variables.}

\item{names}{Vector containing the names of the variables considered in the model,
indicating the scope of the optimization.}

\item{Vi}{Matrix specifying variable pairs, directing the optimization towards relevant
variable interactions.}

\item{uh}{Matrix that combines spatial and temporal distances with additional identifiers,
essential for defining the context of each data point in spatio-temporal space.}

\item{cr}{Initial correlation matrix offering a preliminary estimate of relationships
between variables, serving as a foundation for optimization.}

\item{max_it}{Specifies the maximum number of iterations for the optimization process,
setting a limit to ensure computational feasibility.}

\item{ep}{Data frame defining pairs of variables for detailed analysis, guiding the optimization
process by highlighting specific interactions of interest.}
}
\value{
Updated `par_all` vector containing optimized parameters for each variable pair,
        representing an enhanced set of model parameters post-optimization.
        This updated parameter set is ready for subsequent model application or further analysis.
}
\description{
Optimizes spatio-temporal model parameters for each pair of variables to enhance the
log-likelihood of the observed data under the model. This function focuses on both spatial
and temporal interactions between pairs of variables, refining the model's ability to capture
complex spatio-temporal dependencies.
}
\details{
This function implements the methods described in Section 3.3 of the article, functions log_lik and log
*Stochastic Environmental Research and Risk Assessment, 2025* (DOI: 10.1007/s00477-024-02897-8).
}
\keyword{internal}
