% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRF_helper_functions.R
\name{optimize_temporal_parameters}
\alias{optimize_temporal_parameters}
\title{Optimize Temporal Parameters Across All Variable Pairs}
\usage{
optimize_temporal_parameters(par_all, data, names, Vi, uh, cr, max_it, ep)
}
\arguments{
\item{par_all}{Comprehensive set of model parameters, including both spatial and temporal
parameters, to be optimized in this step.}

\item{data}{Dataset containing spatio-temporal observations, used to evaluate the model's
performance and guide the optimization process.}

\item{names}{Vector of variable names involved in the model, indicating the scope of the
optimization across variable interactions.}

\item{Vi}{Matrix indicating the pairs of variables considered in the model, focusing the
optimization on relevant interactions.}

\item{uh}{Matrix combining spatial and temporal distances along with additional identifiers,
critical for contextualizing each observation in spatio-temporal analysis.}

\item{cr}{Initial correlation matrix providing a baseline of variable relationships, serving
as a starting point for optimization.}

\item{max_it}{Maximum number of iterations allowed in the optimization process, setting
computational bounds to ensure completion.}

\item{ep}{Data frame specifying pairs of variables to be analyzed, guiding the optimization
towards significant interactions.}
}
\value{
Updated `par_all` vector with optimized temporal parameters, representing an
        improved set of model parameters after the optimization process. This vector is
        essential for applying the model to new data or further analysis.
}
\description{
Performs a final optimization step to refine the temporal parameters of the model, 
considering interactions across all variable pairs. This function aims to enhance the 
model's temporal dynamics by optimizing a subset of parameters that influence temporal 
relationships.
}
\keyword{internal}
