% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariance_and_loglikelihood.R
\name{param}
\alias{param}
\title{Construct Covariance Parameters DataFrame}
\usage{
param(par, names)
}
\arguments{
\item{par}{Named vector of parameters.}

\item{names}{Character vector of variable names.}
}
\value{
A data frame where each row corresponds to a pair of variables (including self-pairs) 
and their associated spatio-temporal covariance parameters.
}
\description{
Creates a data frame of covariance parameters for all possible pairs of variables. This function is 
designed for internal use, facilitating the organization of parameters for spatio-temporal modeling.
}
\keyword{internal}
