% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{predict_binomial}
\alias{predict_binomial}
\title{Predict Binomial}
\usage{
predict_binomial(fit, newdata)
}
\arguments{
\item{fit}{Fitted model object containing coefficients.}

\item{newdata}{New data for prediction.}
}
\value{
Predicted binomial probabilities.
}
\description{
Function to predict binomial probabilities using a fitted model object.
}
\keyword{internal}
