% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{rescale_data}
\alias{rescale_data}
\title{Rescale Simulated Data}
\usage{
rescale_data(sim, parm, names, dates)
}
\arguments{
\item{sim}{Simulated weather data, organized as a 3D array where the dimensions are [time, locations, variables].}

\item{parm}{Parameters object containing model parameters, including scale parameters (mu and sd) for each variable.}

\item{names}{Names of the variables included in the simulation.}

\item{dates}{Vector of dates corresponding to the time dimension in the sim array.}
}
\value{
Simulated data with variables other than "Precipitation" rescaled back to their original scale.
}
\description{
Function to rescale simulated weather data for variables other than "Precipitation" to their original scale.
}
\keyword{internal}
