% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_helper_functions.R
\name{season_indices}
\alias{season_indices}
\title{Season Indices Calculation}
\usage{
season_indices(dates, season, Year)
}
\arguments{
\item{dates}{A vector of dates to evaluate.}

\item{season}{A list specifying the season with 'min_day', 'max_day', 'min_month', and 'max_month'.}

\item{Year}{Optional. If specified, only returns indices for the given year.}
}
\value{
A vector of indices corresponding to dates within the specified season.
        If 'Year' is specified, only returns indices for that year.
}
\description{
Calculates the indices of dates within a specified season across multiple years.
This function handles leap years and seasons that span the end of one year and the beginning of the next.
}
\keyword{internal}
