% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_helper_functions.R
\name{simulate_Z}
\alias{simulate_Z}
\title{Simulate AR Process}
\usage{
simulate_Z(Bk, M, num_steps, Z_initial, wt)
}
\arguments{
\item{Bk}{A list of Bk matrices for the AR model.}

\item{M}{The lag order of the AR model.}

\item{num_steps}{The number of time steps to simulate.}

\item{Z_initial}{Initial conditions for the AR process.}

\item{wt}{A vector of weather types corresponding to each time step.}
}
\value{
A list containing the simulated values of Z for each time step.
}
\description{
Simulates the autoregressive (AR) process to generate synthetic weather data.
}
\keyword{internal}
